+++ cypress/e2e/log-in.cy.js
@@ -5,15 +5,13 @@ describe("Log in", () => {
     cy.get('[aria-label="Profile"]').click();
     cy.contains('[role="menuitem"]', "Login").click();
 
-    cy.get('input[value="webId"]').click();
-
     cy.get('input[name="idp"]')
       .clear();
     cy.get('input[name="idp"]')
       .type("http://localhost:8080/example2/profile/card#me");
     cy.contains("Login").click();
 
-    cy.contains("No IDP found");
+    cy.contains("Login failed");
   });
 
   it("Log in with invalid WebID document", () => {
@@ -22,15 +20,33 @@ describe("Log in", () => {
     cy.get('[aria-label="Profile"]').click();
     cy.contains('[role="menuitem"]', "Login").click();
 
-    cy.get('input[value="webId"]').click();
-
     cy.get('input[name="idp"]')
       .clear();
     cy.get('input[name="idp"]')
       .type("http://localhost:8080/invalidWebId/profile/card#me");
     cy.contains("Login").click();
 
-    cy.contains("Couldn't query the Identity Provider from the WebID");
+    cy.contains("Login failed");
+  });
+
+  it("Log in with WebID with OIDC issuer", () => {
+    cy.visit("/");
+
+    cy.get('[aria-label="Profile"]').click();
+    cy.contains('[role="menuitem"]', "Login").click();
+
+    cy.get('input[name="idp"]')
+      .clear();
+    cy.get('input[name="idp"]')
+      .type("http://localhost:8080/example/profile/card#me");
+    cy.contains("Login").click();
+
+    cy.get("input#email").type("hello@example.com");
+    cy.get("input#password").type("abc123");
+    cy.contains("button", "Log in").click();
+    cy.contains("button", "Authorize").click();
+
+    cy.url().should("eq", "http://localhost:5173/");
   });
 
   it("Log in with an invalid IDP issuer", () => {
@@ -48,7 +64,7 @@ describe("Log in", () => {
     cy.contains("Login failed");
   });
 
-  it("Log in and execute query on private data, then log out", () => {
+  it("Log in with a valid IDP issuer and execute query on private data, then log out", () => {
     cy.visit("/");
 
     cy.get('[aria-label="Profile"]').click();